refactor: extract gravity helper
