+++ src/11-brushy-cli/data.ts
@@ -1,3 +1,4 @@
+import { gravity } from "system/gravity"
 import { metaphysics, extractNodes } from "system/metaphysics"
 
 export async function getData({
@@ -171,33 +172,4 @@ export async function getPartnerAppliedGene(name: string) {
 
   const gene = allPartnerAppliedGenes.find((gene) => gene.name === name)
   return gene
-}
-
-export async function gravity(
-  path: string,
-  params: Record<string, unknown> = {},
-  headers: Record<string, string> = {}
-) {
-  const url = new URL(
-    `${process.env.GRAVITY_URL ?? "https://stagingapi.artsy.net"}/api/v1/${path}`
-  )
-  url.search = new URLSearchParams(params as Record<string, string>).toString()
-
-  // console.log(url.toString(), headers)
-
-  const response = await fetch(url.toString(), {
-    headers: {
-      "Content-type": "application/json",
-      ...headers,
-    },
-  })
-
-  if (!response.ok) {
-    const { error, text } = await response.json()
-    throw new Error(
-      `${response.status} (${response.statusText}) -- ${error}: ${text}`
-    )
-  }
-
-  return response.json()
 }
